+++ src/components/pages/ProductListPage.tsx
@@ -1,4 +1,4 @@
-import Header from "../Header/Header";
+import CartHeader from "../Header/CartHeader";
 import ErrorToast from "../Toast/ErrorToast";
 import ProductContent from "../Product/ProductContent";
 import S from "./ProductListPage.module.css";
@@ -16,7 +16,7 @@ const ProductListPage = () => {
 	return (
 		<StoreProvider>
 			<div className={S.container}>
-				<Header onClick={openModal} />
+				<CartHeader onClick={openModal} />
 				<ErrorToast />
 				<ProductContent />
 				<CartModal isOpen={isOpen} setIsOpen={setIsOpen} />
